some changes

+++ src/main/java/mod/maxbogomol/wizards_reborn/common/block/arcane_hopper/ArcaneHopperInventoryCodeHooks.java
@@ -22,28 +22,22 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Optional;
 
 public class ArcaneHopperInventoryCodeHooks {
+
     @Nullable
-    public static Boolean extractHook(Level level, Hopper dest)
-    {
-        return getItemHandler(level, dest, Direction.UP)
-                .map(itemHandlerResult -> {
+    public static Boolean extractHook(Level level, Hopper dest) {
+        return getItemHandler(level, dest, Direction.UP).map(itemHandlerResult -> {
                     IItemHandler handler = itemHandlerResult.getKey();
 
-                    for (int i = 0; i < handler.getSlots(); i++)
-                    {
+                    for (int i = 0; i < handler.getSlots(); i++) {
                         ItemStack extractItem = handler.extractItem(i, 1, true);
-                        if (!extractItem.isEmpty())
-                        {
-                            for (int j = 0; j < dest.getContainerSize(); j++)
-                            {
+                        if (!extractItem.isEmpty()) {
+                            for (int j = 0; j < dest.getContainerSize(); j++) {
                                 ItemStack destStack = dest.getItem(j);
-                                if (dest.canPlaceItem(j, extractItem) && (destStack.isEmpty() || destStack.getCount() < destStack.getMaxStackSize() && destStack.getCount() < dest.getMaxStackSize() && ItemHandlerHelper.canItemStacksStack(extractItem, destStack)))
-                                {
+                                if (dest.canPlaceItem(j, extractItem) && (destStack.isEmpty() || destStack.getCount() < destStack.getMaxStackSize() && destStack.getCount() < dest.getMaxStackSize() && ItemHandlerHelper.canItemStacksStack(extractItem, destStack))) {
                                     extractItem = handler.extractItem(i, 1, false);
                                     if (destStack.isEmpty())
                                         dest.setItem(j, extractItem);
-                                    else
-                                    {
+                                    else {
                                         destStack.grow(1);
                                         dest.setItem(j, destStack);
                                     }
@@ -55,60 +49,45 @@ public class ArcaneHopperInventoryCodeHooks {
                     }
 
                     return false;
-                })
-                .orElse(null);
+                }).orElse(null);
     }
 
-    public static boolean dropperInsertHook(Level level, BlockPos pos, DispenserBlockEntity dropper, int slot, @NotNull ItemStack stack)
-    {
+    public static boolean dropperInsertHook(Level level, BlockPos pos, DispenserBlockEntity dropper, int slot, @NotNull ItemStack stack) {
         Direction enumfacing = level.getBlockState(pos).getValue(DropperBlock.FACING);
         BlockPos blockpos = pos.relative(enumfacing);
-        return getItemHandler(level, (double) blockpos.getX(), (double) blockpos.getY(), (double) blockpos.getZ(), enumfacing.getOpposite())
-                .map(destinationResult -> {
+        return getItemHandler(level, blockpos.getX(), blockpos.getY(), blockpos.getZ(), enumfacing.getOpposite()).map(destinationResult -> {
                     IItemHandler itemHandler = destinationResult.getKey();
                     Object destination = destinationResult.getValue();
                     ItemStack dispensedStack = stack.copy().split(1);
                     ItemStack remainder = putStackInInventoryAllSlots(dropper, destination, itemHandler, dispensedStack);
 
-                    if (remainder.isEmpty())
-                    {
+                    if (remainder.isEmpty()) {
                         remainder = stack.copy();
                         remainder.shrink(1);
-                    }
-                    else
-                    {
+                    } else {
                         remainder = stack.copy();
                     }
 
                     dropper.setItem(slot, remainder);
                     return false;
-                })
-                .orElse(true);
+                }).orElse(true);
     }
 
-    public static boolean insertHook(ArcaneHopperBlockEntity hopper)
-    {
+    public static boolean insertHook(ArcaneHopperBlockEntity hopper) {
         Direction hopperFacing = hopper.getBlockState().getValue(HopperBlock.FACING);
-        return getItemHandler(hopper.getLevel(), hopper, hopperFacing)
-                .map(destinationResult -> {
+        return getItemHandler(hopper.getLevel(), hopper, hopperFacing).map(destinationResult -> {
                     IItemHandler itemHandler = destinationResult.getKey();
                     Object destination = destinationResult.getValue();
-                    if (isFull(itemHandler))
-                    {
+                    if (isFull(itemHandler)) {
                         return false;
-                    }
-                    else
-                    {
-                        for (int i = 0; i < hopper.getContainerSize(); ++i)
-                        {
-                            if (!hopper.getItem(i).isEmpty())
-                            {
+                    } else {
+                        for (int i = 0; i < hopper.getContainerSize(); ++i) {
+                            if (!hopper.getItem(i).isEmpty()) {
                                 ItemStack originalSlotContents = hopper.getItem(i).copy();
                                 ItemStack insertStack = hopper.removeItem(i, 1);
                                 ItemStack remainder = putStackInInventoryAllSlots(hopper, destination, itemHandler, insertStack);
 
-                                if (remainder.isEmpty())
-                                {
+                                if (remainder.isEmpty()) {
                                     return true;
                                 }
 
@@ -118,54 +97,41 @@ public class ArcaneHopperInventoryCodeHooks {
 
                         return false;
                     }
-                })
-                .orElse(false);
+                }).orElse(false);
     }
 
-    private static ItemStack putStackInInventoryAllSlots(BlockEntity source, Object destination, IItemHandler destInventory, ItemStack stack)
-    {
-        for (int slot = 0; slot < destInventory.getSlots() && !stack.isEmpty(); slot++)
-        {
+    private static ItemStack putStackInInventoryAllSlots(BlockEntity source, Object destination, IItemHandler destInventory, ItemStack stack) {
+        for (int slot = 0; slot < destInventory.getSlots() && !stack.isEmpty(); slot++) {
             stack = insertStack(source, destination, destInventory, stack, slot);
         }
         return stack;
     }
 
-    private static ItemStack insertStack(BlockEntity source, Object destination, IItemHandler destInventory, ItemStack stack, int slot)
-    {
+    private static ItemStack insertStack(BlockEntity source, Object destination, IItemHandler destInventory, ItemStack stack, int slot) {
         ItemStack itemstack = destInventory.getStackInSlot(slot);
 
-        if (destInventory.insertItem(slot, stack, true).isEmpty())
-        {
+        if (destInventory.insertItem(slot, stack, true).isEmpty()) {
             boolean insertedItem = false;
             boolean inventoryWasEmpty = isEmpty(destInventory);
 
-            if (itemstack.isEmpty())
-            {
+            if (itemstack.isEmpty()) {
                 destInventory.insertItem(slot, stack, false);
                 stack = ItemStack.EMPTY;
                 insertedItem = true;
-            }
-            else if (ItemHandlerHelper.canItemStacksStack(itemstack, stack))
-            {
+            } else if (ItemHandlerHelper.canItemStacksStack(itemstack, stack)) {
                 int originalSize = stack.getCount();
                 stack = destInventory.insertItem(slot, stack, false);
                 insertedItem = originalSize < stack.getCount();
             }
 
-            if (insertedItem)
-            {
-                if (inventoryWasEmpty && destination instanceof HopperBlockEntity)
-                {
+            if (insertedItem) {
+                if (inventoryWasEmpty && destination instanceof HopperBlockEntity) {
                     HopperBlockEntity destinationHopper = (HopperBlockEntity)destination;
 
-                    if (!destinationHopper.isOnCustomCooldown())
-                    {
+                    if (!destinationHopper.isOnCustomCooldown()) {
                         int k = 0;
-                        if (source instanceof HopperBlockEntity)
-                        {
-                            if (destinationHopper.getLastUpdateTime() >= ((HopperBlockEntity) source).getLastUpdateTime())
-                            {
+                        if (source instanceof HopperBlockEntity) {
+                            if (destinationHopper.getLastUpdateTime() >= ((HopperBlockEntity) source).getLastUpdateTime()) {
                                 k = 1;
                             }
                         }
@@ -178,55 +144,44 @@ public class ArcaneHopperInventoryCodeHooks {
         return stack;
     }
 
-    private static Optional<Pair<IItemHandler, Object>> getItemHandler(Level level, Hopper hopper, Direction hopperFacing)
-    {
+    private static Optional<Pair<IItemHandler, Object>> getItemHandler(Level level, Hopper hopper, Direction hopperFacing) {
         double x = hopper.getLevelX() + (double) hopperFacing.getStepX();
         double y = hopper.getLevelY() + (double) hopperFacing.getStepY();
         double z = hopper.getLevelZ() + (double) hopperFacing.getStepZ();
         return getItemHandler(level, x, y, z, hopperFacing.getOpposite());
     }
 
-    private static boolean isFull(IItemHandler itemHandler)
-    {
-        for (int slot = 0; slot < itemHandler.getSlots(); slot++)
-        {
+    private static boolean isFull(IItemHandler itemHandler) {
+        for (int slot = 0; slot < itemHandler.getSlots(); slot++) {
             ItemStack stackInSlot = itemHandler.getStackInSlot(slot);
-            if (stackInSlot.isEmpty() || stackInSlot.getCount() < itemHandler.getSlotLimit(slot))
-            {
+            if (stackInSlot.isEmpty() || stackInSlot.getCount() < itemHandler.getSlotLimit(slot)) {
                 return false;
             }
         }
         return true;
     }
 
-    private static boolean isEmpty(IItemHandler itemHandler)
-    {
-        for (int slot = 0; slot < itemHandler.getSlots(); slot++)
-        {
+    private static boolean isEmpty(IItemHandler itemHandler) {
+        for (int slot = 0; slot < itemHandler.getSlots(); slot++) {
             ItemStack stackInSlot = itemHandler.getStackInSlot(slot);
-            if (stackInSlot.getCount() > 0)
-            {
+            if (stackInSlot.getCount() > 0) {
                 return false;
             }
         }
         return true;
     }
 
-    public static Optional<Pair<IItemHandler, Object>> getItemHandler(Level level, double x, double y, double z, final Direction side)
-    {
+    public static Optional<Pair<IItemHandler, Object>> getItemHandler(Level level, double x, double y, double z, final Direction side) {
         int i = Mth.floor(x);
         int j = Mth.floor(y);
         int k = Mth.floor(z);
         BlockPos blockpos = new BlockPos(i, j, k);
         net.minecraft.world.level.block.state.BlockState state = level.getBlockState(blockpos);
 
-        if (state.hasBlockEntity())
-        {
+        if (state.hasBlockEntity()) {
             BlockEntity blockEntity = level.getBlockEntity(blockpos);
-            if (blockEntity != null)
-            {
-                return blockEntity.getCapability(ForgeCapabilities.ITEM_HANDLER, side)
-                        .map(capability -> ImmutablePair.<IItemHandler, Object>of(capability, blockEntity));
+            if (blockEntity != null) {
+                return blockEntity.getCapability(ForgeCapabilities.ITEM_HANDLER, side).map(capability -> ImmutablePair.<IItemHandler, Object>of(capability, blockEntity));
             }
         }
